Getting unique pidps for inflated population only in starting years for now - 2014 and 2020. Also tidying some data types up

+++ minos/modules/loneliness.py
@@ -94,7 +94,7 @@ class Loneliness(Base):
                                                               loneliness_prob_df) + 1
         loneliness_prob_df.index = pop.index
 
-        self.population_view.update(loneliness_prob_df["loneliness"])
+        self.population_view.update(loneliness_prob_df["loneliness"].astype(int))
 
     def calculate_loneliness(self, pop):
         """Calculate loneliness transition distribution based on provided people/indices.
